rename facade to manager
