+++ config/sources/stringdb/protein-links-detailed.py
@@ -1,7 +1,7 @@
 from koza.model import Gene, PairwiseGeneToGeneInteraction, predicate
-from koza.facade.data_provider import inject_files, inject_translation_table
-from koza.facade.data_collector import serialize
-from koza.facade.helper import next_row
+from koza.manager.data_provider import inject_files, inject_translation_table
+from koza.manager.data_collector import serialize
+from koza.manager.helper import next_row
 
 _ingest_name = 'protein-links-detailed'
