--- config/translationtable/ctd.yaml
@@ -1,7 +1,0 @@
----
-
-"therapeutic": "is substance that treats"
-"marker/mechanism": "is marker for"
-"signal transduction": "signal_transduction"
-"TAS": "traceable author statement"
-# "pathway": "pathway"  # already exists in globaltt
