--- config/translationtable/decipher.yaml
@@ -1,11 +1,0 @@
----
-
-"Loss of function": "loss_of_function_variant"
-"All missense/in frame": "missense_variant"
-"Dominant negative": "dominant_negative_variant"
-"Activating": "gain_of_function_variant"
-"Increased gene dosage": "copy_number_gain"
-"regulatory region": "regulatory_region_variant"
-"Cis-regulatory or promotor mutation": "regulatory_region_variant"
-"Uncertain": "sequence_variant"
-"5 or 3UTR mutation": "UTR_variant"
