--- config/translationtable/eom.yaml
@@ -1,3 +1,0 @@
----
-# eom.yaml
-"": ""  # example
