--- config/translationtable/genereviews.yaml
@@ -1,3 +1,0 @@
----
-# genereviews.yaml
-"": ""  # example
