--- config/translationtable/hgnc.yaml
@@ -1,37 +1,0 @@
----
-# hgnc.yaml
-
-# locus_type_map
-
-"RNA, Y": "Y RNA"
-"RNA, cluster": "ncRNA"
-"RNA, long non-coding": "lnc_RNA"
-"RNA, micro": "miRNA_gene"
-"RNA, misc": "ncRNA"
-"RNA, ribosomal": "large_subunit_rRNA"       # is this legit? TEC
-"RNA, small cytoplasmic": "scRNA_gene"
-"RNA, small nuclear": "snRNA_gene"
-"RNA, small nucleolar": "snoRNA_gene"
-"RNA, transfer": "tRNA_gene"
-"RNA, vault": "vault_RNA"
-"T cell receptor gene": "vertebrate_immunoglobulin_T_cell_receptor_segment"
-"T cell receptor pseudogene": "T_cell_receptor_pseudogene"
-"complex locus constituent": "protein_coding_gene"  # no way to say member of cluster
-"endogenous retrovirus": "endogenous_retroviral_gene"
-
-# TODO https://sourceforge.net/p/song/term-tracker/433/
-# issue moved to:
-# https://github.com/The-Sequence-Ontology/SO-Ontologies/issues/301
-# where it sits.
-# However: "Fragile Site": "NCIT:C13375"   does exist
-"fragile site": "biological_region"
-"gene with protein product": "protein_coding_gene"
-"immunoglobulin gene": "immunoglobulin_gene"
-"immunoglobulin pseudogene": "immunoglobulin_pseudogene"
-"phenotype only": "heritable_phenotypic_marker"
-"protocadherin": "sequence_feature"        # FIXME  "protocadherin": "PR:000044515" ?
-"readthrough": "stop codon readthrough"
-"transposable element": "transposable_element"
-"unknown": "sequence_feature"
-"virus integration site": "integration_excision_site"  # FIXME better? still not specific
-# "withdrawn": None                          # TODO  NCIT:C38061 ?
