--- config/translationtable/impc.yaml
@@ -1,78 +1,0 @@
----
-# impc.yaml
-"not_applicable": "indeterminate"
-"homozygote": "homozygous"
-"hemizygote": "hemizygous"
-"heterozygote": "simple heterozygous"
-"male": "male intrinsic genotype"
-"female": "female intrinsic genotype"
-"both": "sex_qualified_genotype"
-"no_data": "intrinsic genotype"
-"no data": "intrinsic genotype"
-"not_considered": "unspecified"
-
-
-# project full names no longer supplied as of V12
-# "DTCC consortium": "http://www.criver.com/promo/dtcc"
-# "European Mouse Disease Clinic": "http://www.eumodic.org/"
-# "Helmholtz German Mouse Clinic": "https://www.mouseclinic.de/"
-# "Jackson Laboratory": "https://www.jax.org/"
-# "KOMP2 BaSH consortium": "https://www.mousephenotype.org/"
-# "Model Animal Research Center": "http://www.en.nicemice.cn/"
-# "MRC project": "https://www.har.mrc.ac.uk/"
-# "North American Conditional Mouse Mutagenesis project 2": "http://www.norcomm.org/index.htm"
-# "Phenomin": "http://www.phenomin.fr/"
-# "RIKEN BioResource Center": "http://en.brc.riken.jp/"
-# "RIKEN BioResource Research Center": "http://en.brc.riken.jp/"
-# "Wellcome Trust Sanger Institute Mouse Genetics Project": "http://www.sanger.ac.uk/science/data/mouse-genomes-project"
-# "Korea Mouse Phenotyping Center": "http://mousephenotype.kr/"
-# "International Mouse Phenotyping Consortium": "http://www.mousephenotype.org/"
-# "Infection and Immunity Immunophenotyping (3i) consortium": "http://www.immunophenotype.org/"
-# "Institute of Molecular Genetics of the ASCR, v.v.i": "https://www.img.cas.cz/en/"
-
-# maybe better as part of curie_map ???
-# consider "FOOBAR-SITE": "http://top.level.edu"
-"BCM": "https://www.bcm.edu/"
-"CMHD": "http://www.cmhd.ca/"
-"HMGU": "https://www.helmholtz-muenchen.de/en/"
-"ICS": "http://www.ics-mci.fr/en/departments/phenotyping/"
-"JAX": "https://www.jax.org/"
-"MRC Harwell": "https://www.har.mrc.ac.uk/"
-"NING": "http://www.en.nicemice.cn/"
-"RBRC": "http://en.brc.riken.jp/"
-"TCP": "http://www.phenogenomics.ca/"
-"UC Davis": "https://mbp.mousebiology.org/"
-"WTSI": "http://www.sanger.ac.uk/"
-"MARC": "http://www.nbri-nju.com/en-us/"
-"KMPC": "http://mousephenotype.kr/"
-"IMPC": "https://www.mousephenotype.org/"
-"CCP-IMG": "https://www.phenogenomics.cz/"
-"NorCOMM2": "http://www.norcomm.org/index.htm"
-"DTCC": "http://www.criver.com/promo/dtcc"
-"EUCOMMToolsCre": "https://www.mousephenotype.org/about-impc/about-ikmc/eucommtools/"
-"EUMODIC": "http://www.eumodic.org/"
-"MGP": "http://www.sanger.ac.uk/science/data/mouse-genomes-project"
-"MRC": "https://www.har.mrc.ac.uk/"
-"BaSH": "https://www.mousephenotype.org/"
-"Phenomin": "http://www.phenomin.fr/"
-"Helmholtz GMC": "https://www.mouseclinic.de/"
-"MGP Legacy": "https://www.har.mrc.ac.uk/acknowledgments/"  # ?
-"Infrafrontier-I3": "http://www.immunophenotype.org/"
-# "MRC,BaSH"  -- split out in ingest
-
-# statistical methods
-"Fisher Exact Test framework": "Fisher's exact test"
-"Mixed Model framework, linear mixed-effects model, equation withoutWeight": "linear mixed model"
-"Mixed Model framework, linear mixed-effects model, equation withWeight": "linear mixed model"
-"Unknown": "statistical model"
-"Mixed Model framework, generalized least squares, equation withoutWeight": "generalized least squares estimation"
-"Mixed Model framework, generalized least squares, equation withWeight": "generalized least squares estimation"
-"Wilcoxon rank sum test with continuity correction": "Mann-Whitney U-test"
-"effect_size": "effect size estimate"
-"percentage_change": "fold change"
-"Supplied as data": "statistical model"
-"Linear Mixed Model framework, LME, including Weight": "linear mixed model"
-"Linear Mixed Model framework, LME, not including Weight": "linear mixed model"
-"Linear Model Using Generalized Least Squares framework, GLS, not including Weight": "generalized least squares estimation"
-"Linear Model Using Generalized Least Squares framework, GLS, including Weight": "generalized least squares estimation"
-"Reference Range Plus Test framework; quantile = 0.95 (Tails probability = 0.025)": "statistical model"
