--- config/translationtable/kegg.yaml
@@ -1,3 +1,0 @@
----
-# kegg.yaml
-"": ""  # example
