--- config/translationtable/mgislim.yaml
@@ -1,3 +1,0 @@
----
-# mgislim.yaml
-"": ""  # example
