--- config/translationtable/mmrrc.yaml
@@ -1,3 +1,0 @@
----
-# mmrrc.yaml
-"": ""  # example
