--- config/translationtable/monarch.yaml
@@ -1,3 +1,0 @@
----
-# monarch.yaml
-"": ""  # example
