--- config/translationtable/mpd.yaml
@@ -1,6 +1,0 @@
----
-# mpd.yaml
-"m": "male"
-"f": "female"
-"fm": "male and female"
-"both": "male and female"
