--- config/translationtable/mychem.yaml
@@ -1,3 +1,0 @@
----
-# mychem.yaml
-"": ""  # example
