--- config/translationtable/reactome.yaml
@@ -1,3 +1,0 @@
----
-# reactome.yaml
-"": ""  # example
