--- config/translationtable/sgd.yaml
@@ -1,3 +1,0 @@
----
-# sgd.yaml
-"": ""  # example
