--- config/translationtable/udp.yaml
@@ -1,3 +1,0 @@
----
-# udp.yaml
-"": ""  # example
