--- config/translationtable/zfinslim.yaml
@@ -1,3 +1,0 @@
----
-# zfinslim.yaml
-"": ""  # example
